fix(websocket): preserve architecture boundary
