+++ src/providers/providerLogging.ts
@@ -1,0 +1,21 @@
+import { REDACTED, sanitizeObject, sanitizeUrl } from '../util/sanitizer';
+
+/**
+ * Returns a provider identifier that preserves ordinary URLs while ensuring
+ * credential-bearing URLs are safe to log or persist.
+ */
+export function getSafeProviderId(url: string): string {
+  const sanitizedUrl = sanitizeUrl(url);
+  return sanitizedUrl === REDACTED ||
+    sanitizedUrl.includes(encodeURIComponent(REDACTED)) ||
+    sanitizedUrl.includes('***')
+    ? sanitizedUrl
+    : url;
+}
+
+/**
+ * Sanitizes provider-owned data before including it in logs or error messages.
+ */
+export function sanitizeProviderObject(value: unknown, context: string): unknown {
+  return sanitizeObject(value, { context });
+}
